pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "trader"
version = "2.0.0"
description = "Automated trading analysis and signal system"
requires-python = ">=3.11"
dependencies = [
    "yfinance>=0.2.36",
    "pandas>=2.2.0",
    "ta>=0.11.0",
    "matplotlib>=3.8.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "gspread>=6.0.0",
    "google-auth>=2.28.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "typer>=0.9",
    "structlog>=24.0",
    "tenacity>=8.0",
    "cachetools>=5.0",
]

[project.optional-dependencies]
ml = [
    "prophet>=1.1.5",
    "lightgbm>=4.0",
    "scikit-learn>=1.4",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "mypy>=1.8",
    "ruff>=0.2",
    "black>=24.0",
    "isort>=5.13",
    "pre-commit>=3.6",
]

[project.scripts]
trader = "src.cli:app"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src --cov-fail-under=80"

[tool.isort]
profile = "black"
